fix close buttons conflicts, ensure dvpn works without persistence, make mac openssl generate dvpn certs properly, mac formats USB for ubuntu properly, mac queries network for AKT nodes prooerly, mac does ssh-copy-id to manually chosen akt nodes properly. bump version 0.2.3

+++ aktAPI/Wallet.js
@@ -503,7 +503,9 @@ export class Wallet{
 			fs.writeFileSync(autostartFile,JSON.stringify(params),'utf8');
 		}
 		else{
-			fs.unlinkSync(autostartFile);
+			if(fs.existsSync(autostartFile)){
+				fs.unlinkSync(autostartFile);
+			}
 		}
 	}
 	startProvider(params){
